sdm845: enable bootctrol for db845c build
so that it will be possible to work with GSI image

Test: boot to homescreen
      console:/ # getprop init.svc.vendor.boot-hal-1-1
      running
      console:/ #
      VtsHalBootV1_0TargetTest and VtsHalBootV1_1TargetTest

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>
Change-Id: I147f04de1c8bc7fd506d351f514498341ef7388c

db845c: Update PRODUCT_SHIPPING_API_LEVEL to 31 and FCM to 6
Update PRODUCT_SHIPPING_API_LEVEL and FCM for db845c so they are
at 31 and 6 respectively.

This also dropps the dummy memtrack HAL, as its no longer required.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I96fb44dbb54d8f51dee917175342ee734c074512